+++ BrintDemo/BrinDemoGoldListViewController.h
@@ -10,7 +10,7 @@
 
 @interface BrinDemoGoldListViewController : UITableViewController
 
-@property (nonatomic) int selectedIndex;
-@property (strong, nonatomic) NSIndexPath *selectedRow;
+@property(nonatomic) NSInteger selectedIndex;
+@property(strong, nonatomic) NSIndexPath *selectedRow;
 
 @end
